+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -9,97 +9,239 @@ import net.runelite.client.config.ConfigSection;
 
 @ConfigGroup("dinkplugin")
 public interface DinkPluginConfig extends Config {
+
+    @ConfigSection(
+        name = "Webhook Overrides",
+        description = "Allows webhook data to be sent to a different URL, for the various notifiers",
+        position = -20,
+        closedByDefault = true
+    )
+    String webhookSection = "Webhook Overrides";
+
     @ConfigSection(
         name = "Collection Log",
         description = "Settings for notifying about collection log",
-        position = 1
+        position = 0
     )
     String collectionSection = "Collection Log";
 
     @ConfigSection(
         name = "Pet",
         description = "Settings for notifying when obtaining a pet",
-        position = 4
+        position = 10
     )
     String petSection = "Pet";
 
     @ConfigSection(
         name = "Levels",
         description = "Settings for notifying when levelling a skill",
-        position = 7
+        position = 20
     )
     String levelSection = "Levels";
 
     @ConfigSection(
         name = "Loot",
         description = "Settings for notifying when loot is dropped",
-        position = 11
+        position = 30
     )
     String lootSection = "Loot";
 
     @ConfigSection(
         name = "Death",
         description = "Settings for notifying when you die",
-        position = 16
+        position = 40
     )
     String deathSection = "Death";
 
     @ConfigSection(
         name = "Slayer",
         description = "Settings for notifying when you complete a slayer task",
-        position = 21
+        position = 50
     )
     String slayerSection = "Slayer";
 
     @ConfigSection(
         name = "Quests",
         description = "Settings for notifying when you complete a quest",
-        position = 24
+        position = 60
     )
     String questSection = "Quests";
 
     @ConfigSection(
         name = "Clue Scrolls",
         description = "Settings for notifying when you complete a clue scroll",
-        position = 27
+        position = 70
     )
     String clueSection = "Clue Scrolls";
 
     @ConfigSection(
         name = "Speedruns",
         description = "Settings for notifying when you finish a speedrun",
-        position = 32
+        position = 80
     )
     String speedrunSection = "Speedruns";
 
     @ConfigSection(
         name = "Kill Count",
         description = "Settings for notifying when you kill a boss",
-        position = 37
+        position = 90
     )
     String killCountSection = "Kill Count";
 
     @ConfigSection(
         name = "Combat Tasks",
         description = "Settings for notifying when you complete a combat achievement",
-        position = 43
+        position = 100
     )
     String combatTaskSection = "Combat Tasks";
 
     @ConfigSection(
         name = "Achievement Diary",
         description = "Settings for notifying when you complete an Achievement Diary",
-        position = 47
+        position = 110
     )
     String diarySection = "Achievement Diary";
 
     @ConfigItem(
-        keyName = "discordWebhook",
-        name = "Discord Webhook",
-        description = "The Webhook URL for the discord channel",
-        position = 0
+        keyName = "discordWebhook", // do not rename; would break old configs
+        name = "Primary Webhook URLs",
+        description = "The default webhook URL to send notifications to, if no override is specified.<br/>" +
+            "You can target multiple webhooks by specifying their URLs on separate lines",
+        position = -20
     )
-    default String discordWebhook() {
+    default String primaryWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "collectionWebhook",
+        name = "Collection Webhook Override",
+        description = "If non-empty, collection messages are sent to this URL, instead of the primary URL",
+        position = -19,
+        section = webhookSection
+    )
+    default String collectionWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "petWebhook",
+        name = "Pet Webhook Override",
+        description = "If non-empty, pet messages are sent to this URL, instead of the primary URL",
+        position = -18,
+        section = webhookSection
+    )
+    default String petWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "levelWebhook",
+        name = "Level Webhook Override",
+        description = "If non-empty, level up messages are sent to this URL, instead of the primary URL",
+        position = -17,
+        section = webhookSection
+    )
+    default String levelWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "lootWebhook",
+        name = "Loot Webhook Override",
+        description = "If non-empty, loot messages are sent to this URL, instead of the primary URL",
+        position = -16,
+        section = webhookSection
+    )
+    default String lootWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "deathWebhook",
+        name = "Death Webhook Override",
+        description = "If non-empty, death messages are sent to this URL, instead of the primary URL",
+        position = -15,
+        section = webhookSection
+    )
+    default String deathWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "slayerWebhook",
+        name = "Slayer Webhook Override",
+        description = "If non-empty, slayer messages are sent to this URL, instead of the primary URL",
+        position = -14,
+        section = webhookSection
+    )
+    default String slayerWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "questWebhook",
+        name = "Quest Webhook Override",
+        description = "If non-empty, quest messages are sent to this URL, instead of the primary URL",
+        position = -13,
+        section = webhookSection
+    )
+    default String questWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "clueWebhook",
+        name = "Clue Webhook Override",
+        description = "If non-empty, clue messages are sent to this URL, instead of the primary URL",
+        position = -12,
+        section = webhookSection
+    )
+    default String clueWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "speedrunWebhook",
+        name = "Speedrun Webhook Override",
+        description = "If non-empty, speedrun messages are sent to this URL, instead of the primary URL",
+        position = -11,
+        section = webhookSection
+    )
+    default String speedrunWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "killCountWebhook",
+        name = "Kill Count Webhook Override",
+        description = "If non-empty, kill count messages are sent to this URL, instead of the primary URL",
+        position = -10,
+        section = webhookSection
+    )
+    default String killCountWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "combatTaskWebhook",
+        name = "Combat Task Webhook Override",
+        description = "If non-empty, combat task messages are sent to this URL, instead of the primary URL",
+        position = -9,
+        section = webhookSection
+    )
+    default String combatTaskWebhook() {
+        return "";
+    }
+
+    @ConfigItem(
+        keyName = "diaryWebhook",
+        name = "Diary Webhook Override",
+        description = "If non-empty, achievement diary messages are sent to this URL, instead of the primary URL",
+        position = -8,
+        section = webhookSection
+    )
+    default String diaryWebhook() {
         return "";
     }
 
@@ -141,7 +283,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "petEnabled",
         name = "Enable pets",
         description = "Enable notifications for obtaining pets",
-        position = 4,
+        position = 10,
         section = petSection
     )
     default boolean notifyPet() {
@@ -152,7 +294,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "petSendImage",
         name = "Send Image",
         description = "Send image with the notification",
-        position = 5,
+        position = 11,
         section = petSection
     )
     default boolean petSendImage() {
@@ -163,7 +305,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "petNotifMessage",
         name = "Notification Message",
         description = "The message to be sent through the webhook. Use %USERNAME% to insert your username",
-        position = 6,
+        position = 12,
         section = petSection
     )
     default String petNotifyMessage() {
@@ -174,7 +316,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "levelEnabled",
         name = "Enable level",
         description = "Enable notifications for gaining levels",
-        position = 7,
+        position = 20,
         section = levelSection
     )
     default boolean notifyLevel() {
@@ -185,7 +327,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "levelSendImage",
         name = "Send Image",
         description = "Send image with the notification",
-        position = 8,
+        position = 21,
         section = levelSection
     )
     default boolean levelSendImage() {
@@ -196,7 +338,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "levelInterval",
         name = "Notify Interval",
         description = "Interval between when a notification should be sent",
-        position = 9,
+        position = 22,
         section = levelSection
     )
     default int levelInterval() {
@@ -207,7 +349,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "levelNotifMessage",
         name = "Notification Message",
         description = "The message to be sent through the webhook. Use %USERNAME% to insert your username and %SKILL% to insert the levelled skill(s)",
-        position = 10,
+        position = 23,
         section = levelSection
     )
     default String levelNotifyMessage() {
@@ -218,7 +360,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "lootEnabled",
         name = "Enable loot",
         description = "Enable notifications for gaining loot",
-        position = 11,
+        position = 30,
         section = lootSection
     )
     default boolean notifyLoot() {
@@ -229,7 +371,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "lootSendImage",
         name = "Send Image",
         description = "Send image with the notification",
-        position = 12,
+        position = 31,
         section = lootSection
     )
     default boolean lootSendImage() {
@@ -240,7 +382,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "lootIcons",
         name = "Show loot icons",
         description = "Show icons for the loot obtained",
-        position = 13,
+        position = 32,
         section = lootSection
     )
     default boolean lootIcons() {
@@ -251,7 +393,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "minLootValue",
         name = "Min Loot value",
         description = "Minimum value of the loot to notify",
-        position = 14,
+        position = 33,
         section = lootSection
     )
     default int minLootValue() {
@@ -262,7 +404,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "lootNotifMessage",
         name = "Notification Message",
         description = "The message to be sent through the webhook. Use %USERNAME% to insert your username, %LOOT% to insert the loot and %SOURCE% to show the source of the loot",
-        position = 15,
+        position = 34,
         section = lootSection
     )
     default String lootNotifyMessage() {
@@ -273,7 +415,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "deathEnabled",
         name = "Enable Death",
         description = "Enable notifications for when you die",
-        position = 16,
+        position = 40,
         section = deathSection
     )
     default boolean notifyDeath() {
@@ -284,7 +426,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "deathSendImage",
         name = "Send Image",
         description = "Send image with the notification",
-        position = 17,
+        position = 41,
         section = deathSection
     )
     default boolean deathSendImage() {
@@ -295,7 +437,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "deathNotifMessage",
         name = "Notification Message",
         description = "The message to be sent through the webhook. Use %USERNAME% to insert your username, %VALUELOST% to insert the GE value of the stuff you lost",
-        position = 18,
+        position = 42,
         section = deathSection
     )
     default String deathNotifyMessage() {
@@ -306,7 +448,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "deathNotifPvpEnabled",
         name = "Distinguish PvP deaths",
         description = "Should the plugin use a different message for dying in PvP?",
-        position = 19,
+        position = 43,
         section = deathSection
     )
     default boolean deathNotifPvpEnabled() {
@@ -317,7 +459,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "deathNotifPvpMessage",
         name = "PvP notification message",
         description = "The message to be sent through the webhook. Use %PKER% to insert the killer, %USERNAME% to insert your username, %VALUELOST% to insert the GE value of the stuff you lost",
-        position = 20,
+        position = 44,
         section = deathSection
     )
     default String deathNotifPvpMessage() {
@@ -328,7 +470,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "slayerEnabled",
         name = "Enable Slayer",
         description = "Enable notifications for when you complete a slayer task",
-        position = 21,
+        position = 50,
         section = slayerSection
     )
     default boolean notifySlayer() {
@@ -339,7 +481,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "slayerSendImage",
         name = "Send Image",
         description = "Send image with the notification",
-        position = 22,
+        position = 51,
         section = slayerSection
     )
     default boolean slayerSendImage() {
@@ -350,7 +492,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "slayerPointThreshold",
         name = "Min Slayer Points",
         description = "The minimum slayer task points to warrant a notification",
-        position = 23,
+        position = 52,
         section = slayerSection
     )
     default int slayerPointThreshold() {
@@ -361,7 +503,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "slayerNotifMessage",
         name = "Notification Message",
         description = "The message to be sent through the webhook. Use %USERNAME% to insert your username, %TASK% to insert your task, %POINTS% to show how many points you obtained and %TASKCOUNT% to show how many tasks you have completed.",
-        position = 24,
+        position = 53,
         section = slayerSection
     )
     default String slayerNotifyMessage() {
@@ -372,7 +514,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "questEnabled",
         name = "Enable Quest",
         description = "Enable notifications for when you complete a quest",
-        position = 25,
+        position = 60,
         section = questSection
     )
     default boolean notifyQuest() {
@@ -383,7 +525,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "questSendImage",
         name = "Send Image",
         description = "Send image with the notification",
-        position = 26,
+        position = 61,
         section = questSection
     )
     default boolean questSendImage() {
@@ -394,7 +536,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "questNotifMessage",
         name = "Notification Message",
         description = "The message to be sent through the webhook. Use %USERNAME% to insert your username and %QUEST% to insert the quest that you completed",
-        position = 27,
+        position = 62,
         section = questSection
     )
     default String questNotifyMessage() {
@@ -405,7 +547,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "clueEnabled",
         name = "Enable Clue Scrolls",
         description = "Enable notifications for when you complete a clue scroll",
-        position = 28,
+        position = 70,
         section = clueSection
     )
     default boolean notifyClue() {
@@ -416,7 +558,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "clueSendImage",
         name = "Send Image",
         description = "Send image with the notification",
-        position = 29,
+        position = 71,
         section = clueSection
     )
     default boolean clueSendImage() {
@@ -427,7 +569,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "clueShowItems",
         name = "Show Item Icons",
         description = "Show item icons gained from the clue",
-        position = 30,
+        position = 72,
         section = clueSection
     )
     default boolean clueShowItems() {
@@ -438,7 +580,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "clueMinValue",
         name = "Min Value",
         description = "The minimum value of the items to be shown",
-        position = 31,
+        position = 73,
         section = clueSection
     )
     default int clueMinValue() {
@@ -449,7 +591,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "clueNotifMessage",
         name = "Notification Message",
         description = "The message to be sent through the webhook. Use %USERNAME% to insert your username, %CLUE% to insert the clue type, %LOOT% to show the loot obtained and %COUNT% to insert how many of those clue types you have completed",
-        position = 32,
+        position = 74,
         section = clueSection
     )
     default String clueNotifyMessage() {
@@ -460,7 +602,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "speedrunEnabled",
         name = "Enable speedruns",
         description = "Enable notifications for when you complete a speedrun",
-        position = 33,
+        position = 80,
         section = speedrunSection
     )
     default boolean notifySpeedrun() {
@@ -471,7 +613,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "speedrunSendImage",
         name = "Send Image",
         description = "Send image with the notification",
-        position = 34,
+        position = 81,
         section = speedrunSection
     )
     default boolean speedrunSendImage() {
@@ -482,7 +624,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "speedrunPBOnly",
         name = "Notify on Personal Best only",
         description = "Enable notifications only for your best runs",
-        position = 35,
+        position = 82,
         section = speedrunSection
     )
     default boolean speedrunPBOnly() {
@@ -493,7 +635,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "speedrunPBMessage",
         name = "PB message",
         description = "%USERNAME% to insert your username, %QUEST% to insert the quest name, %TIME% to insert your new time",
-        position = 36,
+        position = 83,
         section = speedrunSection
     )
     default String speedrunPBMessage() {
@@ -504,7 +646,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "speedrunMessage",
         name = "Notification message",
         description = "%USERNAME% to insert your username, %QUEST% to insert the quest name, %TIME% to insert your new time, %BEST% to insert your PB",
-        position = 37,
+        position = 84,
         section = speedrunSection
     )
     default String speedrunMessage() {
@@ -515,7 +657,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "killCountEnabled",
         name = "Enable Kill Count",
         description = "Enable notifications for boss kill count milestones",
-        position = 38,
+        position = 90,
         section = killCountSection
     )
     default boolean notifyKillCount() {
@@ -526,7 +668,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "killCountSendImage",
         name = "Send Image",
         description = "Send image with the notification",
-        position = 39,
+        position = 91,
         section = killCountSection
     )
     default boolean killCountSendImage() {
@@ -537,7 +679,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "killCountInitial",
         name = "Initial Boss Kill",
         description = "Notify on the first ever kill of any boss",
-        position = 40,
+        position = 92,
         section = killCountSection
     )
     default boolean killCountNotifyInitial() {
@@ -548,7 +690,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "killCountInterval",
         name = "Kill Count Interval",
         description = "Interval between when a notification should be sent",
-        position = 41,
+        position = 93,
         section = killCountSection
     )
     default int killCountInterval() {
@@ -559,7 +701,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "killCountMessage",
         name = "Notification Message",
         description = "The message to be sent to the webhook. Use %USERNAME% to insert your username, %BOSS% to insert the NPC name, %COUNT% to insert the kill count",
-        position = 42,
+        position = 94,
         section = killCountSection
     )
     default String killCountMessage() {
@@ -570,7 +712,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "combatTaskEnabled",
         name = "Enable Combat Tasks",
         description = "Enable notifications for combat achievements",
-        position = 43,
+        position = 100,
         section = combatTaskSection
     )
     default boolean notifyCombatTask() {
@@ -581,7 +723,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "combatTaskSendImage",
         name = "Send Image",
         description = "Send image with the notification",
-        position = 44,
+        position = 101,
         section = combatTaskSection
     )
     default boolean combatTaskSendImage() {
@@ -592,7 +734,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "combatTaskMinTier",
         name = "Min Tier",
         description = "Minimum combat achievement tier to warrant a notification",
-        position = 45,
+        position = 102,
         section = combatTaskSection
     )
     default CombatAchievementTier minCombatAchievementTier() {
@@ -603,7 +745,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "combatTaskMessage",
         name = "Notification Message",
         description = "The message to be sent to the webhook. Use %USERNAME% to insert your username, %TIER% to insert the task tier, %TASK% to insert the task name",
-        position = 46,
+        position = 103,
         section = combatTaskSection
     )
     default String combatTaskMessage() {
@@ -614,7 +756,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "diaryEnabled",
         name = "Enable Diary",
         description = "Enable notifications for achievement diary completions",
-        position = 47,
+        position = 110,
         section = diarySection
     )
     default boolean notifyAchievementDiary() {
@@ -625,7 +767,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "diarySendImage",
         name = "Send Image",
         description = "Send image with the notification",
-        position = 48,
+        position = 111,
         section = diarySection
     )
     default boolean diarySendImage() {
@@ -636,7 +778,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "diaryMinDifficulty",
         name = "Min Difficulty",
         description = "Minimum achievement diary difficulty to warrant a notification",
-        position = 49,
+        position = 112,
         section = diarySection
     )
     default AchievementDiaries.Difficulty minDiaryDifficulty() {
@@ -647,7 +789,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "diaryMessage",
         name = "Notification Message",
         description = "The message to be sent to the webhook. Use %USERNAME% to insert your username, %DIFFICULTY% to insert the diary difficulty, %AREA% to insert the diary area, %TOTAL% to insert the total diaries completed",
-        position = 50,
+        position = 113,
         section = diarySection
     )
     default String diaryNotifyMessage() {