+++ src/test/java/dinkplugin/notifiers/DeathNotifierTest.java
@@ -81,6 +81,7 @@ class DeathNotifierTest extends MockedNotifierTest {
 
         // verify notification
         verify(messageHandler).createMessage(
+            PRIMARY_WEBHOOK_URL,
             false,
             NotificationBody.builder()
                 .content(String.format("%s has died, losing %d gp", PLAYER_NAME, 0))
@@ -122,6 +123,7 @@ class DeathNotifierTest extends MockedNotifierTest {
             new NotificationBody.Embed(new NotificationBody.UrlEmbed("https://static.runelite.net/cache/item/icon/" + ItemID.TUNA + ".png"))
         );
         verify(messageHandler).createMessage(
+            PRIMARY_WEBHOOK_URL,
             false,
             NotificationBody.builder()
                 .content(String.format("%s has died, losing %d gp", PLAYER_NAME, TUNA_PRICE))
@@ -146,6 +148,7 @@ class DeathNotifierTest extends MockedNotifierTest {
 
         // verify notification
         verify(messageHandler).createMessage(
+            PRIMARY_WEBHOOK_URL,
             false,
             NotificationBody.builder()
                 .content(String.format("%s has been PKed by %s for %d gp", PLAYER_NAME, pker, 0))
@@ -172,6 +175,7 @@ class DeathNotifierTest extends MockedNotifierTest {
 
         // verify notification
         verify(messageHandler).createMessage(
+            PRIMARY_WEBHOOK_URL,
             false,
             NotificationBody.builder()
                 .content(String.format("%s has been PKed by %s for %d gp", PLAYER_NAME, pker, 0))
@@ -195,6 +199,7 @@ class DeathNotifierTest extends MockedNotifierTest {
 
         // verify non-PK notification
         verify(messageHandler).createMessage(
+            PRIMARY_WEBHOOK_URL,
             false,
             NotificationBody.builder()
                 .content(String.format("%s has died, losing %d gp", PLAYER_NAME, 0))
@@ -213,7 +218,7 @@ class DeathNotifierTest extends MockedNotifierTest {
         notifier.onActorDeath(new ActorDeath(other));
 
         // ensure no notification occurred
-        verify(messageHandler, never()).createMessage(anyBoolean(), any());
+        verify(messageHandler, never()).createMessage(any(), anyBoolean(), any());
     }
 
     @Test
@@ -225,7 +230,7 @@ class DeathNotifierTest extends MockedNotifierTest {
         notifier.onActorDeath(new ActorDeath(localPlayer));
 
         // ensure no notification occurred
-        verify(messageHandler, never()).createMessage(anyBoolean(), any());
+        verify(messageHandler, never()).createMessage(any(), anyBoolean(), any());
     }
 
     @Test
@@ -237,7 +242,7 @@ class DeathNotifierTest extends MockedNotifierTest {
         notifier.onActorDeath(new ActorDeath(localPlayer));
 
         // ensure no notification occurred
-        verify(messageHandler, never()).createMessage(anyBoolean(), any());
+        verify(messageHandler, never()).createMessage(any(), anyBoolean(), any());
     }
 
 }